docs: move to starlight (#627)
* docs: remove from main branch

These will be in the `site` branch from now on

• Removed input 'catppuccin-v1_1'
• Removed input 'catppuccin-v1_2'
• Removed input 'nuscht-search'
• Removed input 'nuscht-search/flake-utils'
• Removed input 'nuscht-search/ixx'
• Removed input 'nuscht-search/ixx/flake-utils'
• Removed input 'nuscht-search/ixx/nixpkgs'
• Removed input 'nuscht-search/nixpkgs'

* ci: use single `nix flake check` invocation

The previous split was largely pointless as the module tests end up
using all of the packages anyways. Lets streamline things

* chore(devflake): export module json options

For use in the website

* fix(home-manager/vscode): add description to accent option

* docs: move to starlight

* style(treewide): run prettier

* ci: auto-format with prettier

* fix(hm/firefox): pointlessly re-declare option information

Maybe doing full eval with home-manager would be good, but then we
couldn't easily evaluate old versions ourself. Womp womp

* chore(devflake): only put packages in checks